requirements.txt
Flask
websocket-client
websockets
requests
python-telegram-bot==20.3
